requirements.txt
ttkbootstrap
Pillow
yt-dlp
